feat: add warp fee vault
