--- .changeset/warp-fee-splitter.md
@@ -1,6 +1,0 @@
----
-'@hyperlane-xyz/core': minor
-'@hyperlane-xyz/sdk': minor
----
-
-Added the warp fee splitter contract and SDK helper for splitting swept warp route fees and streaming LP donations into hub-chain LP routers.
