+++ .changeset/warp-fee-vault.md
@@ -1,0 +1,6 @@
+---
+'@hyperlane-xyz/core': minor
+'@hyperlane-xyz/sdk': minor
+---
+
+Added the warp fee vault contract and SDK helper for wrapping hub-chain LP router positions while streaming swept warp route fees into vault accounting.
